feat: add the Cal function
The function will calculate the values which we will work with in each
iteration. It accept "i", "indexValue" as int pointer(they will point to
the same name variables in the "CanJump" function).
The value which is int will set to the "indexValue" and main logic will
work with it in the "CanJump" function.